feat: allow getting the config based on the otp_app (#17)
In ueberauth 0.6, support was added to allow multiple configurations of a strategy based on the otp_app. This simple patch adds in support for this, allowing different elixir apps to have separate cognito configurations.

This change is backward compatible, and if the otp_app is not set, it fails back to the default :ueberauth app.